HiKey: pll.c: do not access memory at address 0x0
This commit deletes the memory access test done in BL1 immediately after
DDR initialization. The test is performed at address 0x0, which is valid
but may be subject to an unwanted compiler optimization. GCC 4.9, for
instance, assumes that programs cannot safely dereference null pointers
and will generate a brk instruction, resulting in a BL1 hang.

Using -fno-delete-null-pointer-checks would be another option, but it is
not justified because we don't really need to access address 0x0 anyway.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>