service: psci: add mailbox support
When cpu is to be powered off, it will enter "wfi" state and power
controller can power off it safely. This patch just add the mailbox
mechanism for the cpu cannot be powered off on hikey, so that cpu will
exit from wfi state; for this case, can use mailbox for every cpu for
the branch address, so that cpu can continue run to resume back to
system.

Signed-off-by: Leo Yan <leo.yan@linaro.org>