hikey: increase NS BL1U region size
When TBB is enabled in ARM Trusted Firmware, more size is required
on BL1 RO region in ATF, which shifts up the base of NS BL1U.
NS BL1U in ATF is BL1 RO in atf-fastboot, so we need to shift its
base here also. Otherwise, it fails in recovery mode.

Signed-off-by: Teddy Reed <teddy@prosauce.org>
Signed-off-by: Victor Chong <victor.chong@linaro.org>
Signed-off-by: Haojian Zhuang <haojian.zhuang@linaro.org>